requirements.txt
selenium==4.16.0
webdriver-manager==4.0.1
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
python-dotenv==1.0.0
schedule==1.2.0
lxml==4.9.3
